Config box: move legacy protocols to below SSH.
The contribution of SUPDUP has pushed SSH off the first page in my
default GTK font settings. It probably won't do that the same way for
everyone, but it did make me realise that the [Telnet, Rlogin, SSH]
order was a relic of 1997 when the SSH backend was new and added at
the bottom of the list. These days, SSH ought to be at the top for
easy access!

(And Serial ought to be next, but since it's added in a separate file,
that involves a bit more faffing about which I haven't done yet.)